Fixed uint64_t to int conversion warning

+++ libobs/util/windows/custom_patch.h
@@ -23,7 +23,7 @@ struct WINDOW_DATA {
 	DWORD tid;
 	char caption_name[256];
 	char class_name[256];
-	int monitor_id;
+	uint64_t monitor_id;
 	bool primary_monitor;
 	bool visible_on_screen;
 };
